Removed the time from the hash calculation
I tried some workarounds to handle the case where different bots in the
same channel see messages at slightly different times, but there was
always the possibility of a conflict that had to be inspected and
resolved manually.

Rather than accept that, I just removed the time from the hash
calculation altogether. That ensures that two bots will always calculate
the same ID for the same message.